refactor(ts): Task 7 批次 A+B —— 13 个根模块迁 strict TS
批次 A（9 个小模块）+ 批次 B（4 个中等模块）全部自 JS 迁 strict TS，
d.ts 垫片随迁删除；导出面与运行时语义不变：

- error-detail / bundle-integrity / stable-port / session-encoding-heal /
  plugin-manager-state / profile-module-heal / watchdog（detached 进程入口）
  / preset-sync / builtin-collision
- koffi-preflight（依赖注入面类型化；十六进制退出码日志格式保留——测试
  断言依赖）/ session-watcher（zstd 帧扫描 + TurnEndInfo 类型化）/
  balance（峰谷定价 + PriceEntry 联合类型，balance-ui 消费面适配）/
  patch-row-heal（行修复器全家桶）

附带修复：
- 删除遗留备份 .tmp-orig-main.js（4424 行死代码）
- tsconfig include 改为根目录 *.ts 通配（消逐文件枚举）
- session-watcher 头注释的 sessions/**/ 路径通配曾提前终止块注释（**/
  含 */）——改写措辞
- boot.ts bundle-manifest 解析补 BundleManifest 类型
- balance-ui TurnEndInfo.sessionId 放宽为 string | null（通知限频键 ?? ''）

测试：全量 487/487 通过；tsc --noEmit 零错误。
