--- dsh-desktop/plugin-manager-state.d.ts
@@ -1,4 +1,0 @@
-/** plugin-manager-state.d.ts — legacy `plugin-manager-state.js` 垫片。迁 TS 后删除。 */
-export declare function collectPluginRows(
-  entries: unknown[], opts: Record<string, unknown>,
-): unknown[];
