--- dsh-desktop/profile-module-heal.d.ts
@@ -1,4 +1,0 @@
-/** profile-module-heal.d.ts — legacy `profile-module-heal.js` 垫片。迁 TS 后删除。 */
-export declare function healProfileModuleShadowing(
-  home: string, profile?: string, log?: (msg: string) => void,
-): string[];
