+++ dsh-desktop/tsconfig.json
@@ -18,6 +18,6 @@
     "types": ["node"],
     "noEmit": false
   },
-  "include": ["lib/**/*.ts", "shared/**/*.ts", "preload/**/*.ts", "host-bootstrap.ts", "client-updater.ts", "logger.ts", "plugin-guard.ts", "preload.ts", "types-*.d.ts"],
+  "include": ["lib/**/*.ts", "shared/**/*.ts", "preload/**/*.ts", "*.ts", "types-*.d.ts"],
   "exclude": ["node_modules", "vendor", "dist", "out", "assets", "docs"]
 }
