+++ dsh-desktop/watchdog.ts
@@ -1,22 +1,23 @@
-'use strict';
+/**
+ * watchdog.ts — 桌面端看门狗进程（Task 7.1 自 watchdog.js 迁 TS）。
+ *
+ * Electron 主进程在 boot 时以 detached 方式拉起这个微型 Node 进程
+ * （编译产物 watchdog.js，由 node 直接执行）。它轮询父进程 PID：
+ *   - <userData>/run-state.json 里 cleanExit=true → 用户主动退出/更新/
+ *     fatal-boot 路径已标记为有意退出，看门狗静默退场；
+ *   - 状态文件已被更新的实例接管 → 本看门狗退场；
+ *   - 否则应用意外死亡 → 重新拉起 <exe>。
+ *
+ * 护栏：10 分钟内最多重启 5 次；每次拉起后 15s 宽限期（等新实例先写完
+ * 自己的 run-state 文件）。
+ */
 
-// DSH Desktop watchdog: keeps the packaged desktop app alive.
-//
-// The Electron main process launches this tiny Node process detached at boot.
-// It polls the parent PID. If the parent disappears:
-//   - cleanExit=true in <userData>/run-state.json  -> user quit, update, or
-//     fatal-boot path marked the exit intentionally; watchdog exits quietly.
-//   - a NEWER instance already took over the state file -> this watchdog exits.
-//   - otherwise the app died unexpectedly -> relaunch <exe>.
-//
-// Guard rails: at most 5 relaunches per 10 minutes, and a 15s grace period
-// after each launch so the new instance can write its run-state file first.
+import * as fs from 'node:fs';
+import * as path from 'node:path';
+import { spawn } from 'node:child_process';
 
-const fs = require('node:fs');
-const path = require('node:path');
-const { spawn } = require('node:child_process');
-
-function arg(name, fallback) {
+/** 从 argv 取 --name=value 参数（缺失返回 fallback）。 */
+function arg(name: string, fallback: string): string {
   const prefix = '--' + name + '=';
   const hit = process.argv.find((a) => a.startsWith(prefix));
   return hit ? hit.slice(prefix.length) : fallback;
@@ -35,27 +36,41 @@ let restartCount = 0;
 let windowStart = 0;
 let lastLaunchAt = 0;
 
-function log(msg) {
-  if (!logFile) return;
-  const line = `[${new Date().toISOString()}] ${msg}\n`;
-  try { fs.appendFileSync(logFile, line, 'utf8'); } catch {}
+/** run-state.json 的最小形状。 */
+interface RunState {
+  cleanExit?: boolean;
+  pid?: number;
 }
 
-function alive(pid) {
+function log(msg: string): void {
+  if (!logFile) return;
+  const line = `[${new Date().toISOString()}] ${msg}\n`;
+  try {
+    fs.appendFileSync(logFile, line, 'utf8');
+  } catch {
+    /* 日志写失败静默 */
+  }
+}
+
+function alive(pid: number): boolean {
   if (!pid || pid <= 0) return false;
   try {
     process.kill(pid, 0);
     return true;
   } catch (err) {
-    return err && err.code === 'EPERM';
+    return (err as NodeJS.ErrnoException).code === 'EPERM';
   }
 }
 
-function readState() {
-  try { return JSON.parse(fs.readFileSync(stateFile, 'utf8')); } catch { return null; }
+function readState(): RunState | null {
+  try {
+    return JSON.parse(fs.readFileSync(stateFile, 'utf8')) as RunState;
+  } catch {
+    return null;
+  }
 }
 
-function launchApp() {
+function launchApp(): void {
   const now = Date.now();
   if (now - lastLaunchAt < GRACE_MS) return;
   if (restartCount === 0) windowStart = now;
@@ -83,11 +98,11 @@ function launchApp() {
     });
     child.unref();
   } catch (err) {
-    log('watchdog: spawn failed: ' + ((err && err.message) || err));
+    log('watchdog: spawn failed: ' + String((err as Error).message || err));
   }
 }
 
-function poll() {
+function poll(): void {
   if (alive(watchedPid)) return;
   const state = readState();
   if (state && state.cleanExit === true) {
